fix: Avoid Sympy printer rendering modifications
Adds logic to the latex renderer to fix edge cases for expressions with
subscripts. In specific, mathematical parameters with subscripts may
contain expressions that are in the default modifier dictionary of
Sympy's latex printer. This printer is used by default when creating
``Symbol`` objects in Sympy. Therefore, some texts may be rendered in
unexpected ways if naively printed using the Sympy's ``LatexPrinter``.